push

+++ convex/functions.ts
@@ -1,9 +1,23 @@
 import { query, mutation } from "./_generated/server";
 import { v } from "convex/values";
+import { requireUser } from "./helper";
 
 // Query to list todos
-export const listTodos = query(async ({ db }) => {
-  return await db.query("todos").collect();
+// Query to list todos
+export const listTodos = query({
+  handler: async (ctx) => {
+    const user = await ctx.auth.getUserIdentity();
+
+    // If the user is not signed in, return an empty array
+    if (!user) {
+      return [];
+    }
+
+    // If the user is signed in, return their todos
+    return await ctx.db.query("todos")
+      .withIndex("by_userId", q => q.eq("userId", user.tokenIdentifier))
+      .collect();
+  }
 });
 
 // Mutation to create a new todo
@@ -13,10 +27,12 @@ export const createTodo = mutation({
     description: v.string(),
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx); 
     await ctx.db.insert("todos", {
       title: args.title,
       description: args.description,
       completed: false, 
+      userId: user.tokenIdentifier,
     });
   },
 });
@@ -28,6 +44,11 @@ export const updateTodo = mutation({
     completed: v.boolean(),
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx); 
+    const todo = await ctx.db.get(args.id);
+    if (todo?.userId !== user.tokenIdentifier) {
+      throw new Error("Unauthorized")
+    }
     await ctx.db.patch(args.id, {
       completed: args.completed,
     });
@@ -40,6 +61,11 @@ export const deleteTodo = mutation({
     id: v.id("todos"), 
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx); 
+    const todo = await ctx.db.get(args.id);
+    if (todo?.userId !== user.tokenIdentifier) {
+      throw new Error("Unauthorized")
+    }
     await ctx.db.delete(args.id); 
   },
 });
